+++ destinations/airbyte-faros-destination/src/converters/phabricator/revision_diffs.ts
@@ -1,18 +1,10 @@
 import {AirbyteLogger, AirbyteRecord} from 'faros-airbyte-cdk';
-import {uniq} from 'lodash';
+import {FileDiff} from 'faros-airbyte-common/common';
 import {Dictionary} from 'ts-essentials';
 
+import {Common} from '../common/common';
 import {DestinationModel, DestinationRecord} from '../converter';
 import {PhabricatorCommon, PhabricatorConverter} from './common';
-
-interface FileDiff {
-  deletions: number;
-  additions: number;
-  from?: string;
-  to?: string;
-  deleted?: boolean;
-  new?: boolean;
-}
 
 interface RevisionDiff {
   id: number;
@@ -25,8 +17,6 @@ interface RevisionDiff {
   repository: Dictionary<any>;
   files: ReadonlyArray<FileDiff>;
 }
-
-const NULL = '/dev/null';
 
 export class RevisionDiffs extends PhabricatorConverter {
   private readonly logger: AirbyteLogger = new AirbyteLogger();
@@ -74,75 +64,8 @@ export class RevisionDiffs extends PhabricatorConverter {
       uid: diff.revision.id.toString(),
       repository,
     };
-    const filesChanged = uniq(
-      diff.files.flatMap((f) => [f.from, f.to]).filter((f) => f && f !== NULL)
-    );
 
-    res.push({
-      model: 'vcs_PullRequest__Update',
-      record: {
-        at: Date.now(),
-        where: pullRequest,
-        mask: ['diffStats'],
-        patch: {
-          diffStats: {
-            filesChanged: filesChanged.length,
-            linesAdded: diff.files.reduce(
-              (total, file) => total + file.additions,
-              0
-            ),
-            linesDeleted: diff.files.reduce(
-              (total, file) => total + file.deletions,
-              0
-            ),
-          },
-        },
-      },
-    });
-
-    for (const uid of filesChanged) {
-      res.push({
-        model: 'vcs_File',
-        record: {uid, path: uid, repository},
-      });
-    }
-
-    for (const file of diff.files) {
-      if (file.from && file.from !== NULL && file.from === file.to) {
-        res.push({
-          model: 'vcs_PullRequestFile',
-          record: {
-            file: {uid: file.from, path: file.from, repository},
-            pullRequest,
-            additions: file.additions,
-            deletions: file.deletions,
-          },
-        });
-      } else {
-        if (file.from !== NULL) {
-          res.push({
-            model: 'vcs_PullRequestFile',
-            record: {
-              file: {uid: file.from, path: file.from, repository},
-              pullRequest,
-              additions: 0,
-              deletions: file.deletions,
-            },
-          });
-        }
-        if (file.to !== NULL) {
-          res.push({
-            model: 'vcs_PullRequestFile',
-            record: {
-              file: {uid: file.to, path: file.to, repository},
-              pullRequest,
-              additions: file.additions,
-              deletions: 0,
-            },
-          });
-        }
-      }
-    }
+    res.push(...Common.processVcsPullRequestFileDiffs(diff.files, pullRequest));
     return res;
   }
 }
