+++ faros-airbyte-common/src/common/index.ts
@@ -1,0 +1,8 @@
+export interface FileDiff {
+  deletions: number;
+  additions: number;
+  from?: string;
+  to?: string;
+  deleted?: boolean;
+  new?: boolean;
+}
